requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
Werkzeug==3.0.1
keras==2.13.1
tensorflow-cpu
pillow
efficientnet
gunicorn
flask-cors
sqlalchemy
boto3
flask_sqlalchemy